Renaming all ecoevolity configs.
Renaming all ecoevolity configs to reflect their new settings.  The
number of pairs increased from 8 to 10, and the priors on the model
prior parameters were updated accordingly so that the prior mean number
of div time categories is 5.5. I also increased the variance of these
priors so that the prior probability was spread more evenly across the
possible numbers of categories.