Fixing bug in submit_sim_analyses.sh.
For some reason the following call to grep would fail completely
silently and cause the script to exit with any warnings:

    runtime_line_count="$(grep -c "Runtime:" "$out_file")"

When we were using full paths, this worked fine. But, now that the path
of $out_files starts with "../" we are seeing this behavior.

However, if I don't try to set a variable to the output of grep, it
works fine. So now, "$(grep -c "Runtime:" "$out_file")" is used directly
in the if statement (rather than using runtime_line_count), and it seems
to be working.

Weird Bash behavior.