Working on path behavior in ecoevolity qsub scripts.
Now that we will be using job arrays to run all the ecoevolity analyses
of the simulated datasets, I had to overhaul how paths are handled in
the bash scripts that run ecoevolity.  cd'ing into the PBS_O_WORKDIR
didn't work any more, because these bash scripts were being spawned by
'spawn_job_array' script, which was submitted to the queue by 'psub.'

To fix this, I'm udpating project_util.py to make the PBS headers cd to
the path of the bash script that runs ecoevolity.

Also, updating 'submit_sim_analyses_array.sh' script so that the temp
file with paths to all the ecoevolity bash scripts contains absolute
paths. Also, this temp file is now created in the working directory,
because /tmp is not visible from the compute nodes.