Adding skeleton of documentation.
Adding skeleton of the source files for the documentation that will be
compiled using Sphinx.

All of the documentation source files are contained in
docs-source/source. The make file to compile these source files into
html files is docs-source/Makefile. When 'make html' is run from the
docs-source directory, Sphinx compiles the html from these source files
and puts the html files in docs-source/build. I've also added a
'publish' make target, so that when 'make html' is run from the
docs-source directory, the html files are also copied to the 'docs'
directory. These are the files recognized and served on the web by
GitHub.