Updating python and shell scripts.
Updating all scripts to source the bashrc file if job is being run as a
PBS job submission.

This change should not affect any of the results up to this point, but
will hopefully fixe the problem with some Hopper users not being able to
run the python scripts for plotting.