Updating parse_sim_results.py.
Updating parsing script to include summary of model distances, which are
calculated as distances between set partitions using the
Hungarian/Munkres algorithm (this has recently been added to
pycoevolity).

For this script to work, the pycoevolity installation will need
to be updated to include the new code for calculating these
model distances.