Removing batch 508355009 of simulations.
The archives of simulation files were missing for this batch (only the
summary .tsv files were present). Because we need to re-parse the
results files to add model distance summaries to the tsv files, we
cannot do this with this batch. The easiest fix is just to delete it for
now. We can always replace it with another batch of 10 replicates.